requirements.txt
pyjwt>=2.8.0
requests>=2.31.0
